backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
elasticsearch==8.11.0
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.4
openpyxl==3.1.2
stripe==7.4.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
httpx==0.25.2
email-validator==2.1.0
sendgrid==6.11.0
pillow==10.1.0
aiofiles==23.2.1
jinja2==3.1.2
slowapi==0.1.9
